Cleanup, added missing dependencies and Soft Light Shader
