--- picture/pictureSprite.vert
@@ -1,13 +1,0 @@
-attribute vec2 aVertexPosition;
-attribute vec2 aTextureCoord;
-attribute vec4 aColor;
-
-uniform mat3 projectionMatrix;
-
-varying vec2 vTextureCoord;
-
-void main(void)
-{
-    gl_Position = vec4((projectionMatrix * vec3(aVertexPosition, 1.0)).xy, 0.0, 1.0);
-    vTextureCoord = aTextureCoord;
-}
