+++ src/SoftLightShader.js
@@ -6,17 +6,17 @@ var glslify  = require('glslify');
  * @memberof PIXI.tilemap
  * @param gl {PIXI.Shader} The WebGL shader manager this shader works for.
  */
-function HardLightShader(gl)
+function SoftLightShader(gl)
 {
     PIXI.Shader.call(this,
         gl,
         glslify('./blend.vert'),
-        glslify('./hard_light.frag')
+        glslify('./soft_light.frag')
     );
     this.bind();
     this.uniforms.uSampler = [0, 1];
 }
 
-HardLightShader.prototype = Object.create(PIXI.Shader.prototype);
-HardLightShader.prototype.constructor = HardLightShader;
-module.exports = HardLightShader;
+SoftLightShader.prototype = Object.create(PIXI.Shader.prototype);
+SoftLightShader.prototype.constructor = SoftLightShader;
+module.exports = SoftLightShader;
