+++ src/mapFilterBlendModesToPixi.js
@@ -1,6 +1,7 @@
 var CONST = PIXI,
     OverlayShader = require('./OverlayShader'),
     HardLightShader = require('./HardLightShader');
+    SoftLightShader = require('./SoftLightShader');
 
 /**
  * Maps gl blend combinations to WebGL
@@ -15,6 +16,7 @@ function mapFilterBlendModesToPixi(gl, array)
     //add a boolean for that!
     array[CONST.BLEND_MODES.OVERLAY] = new OverlayShader(gl);
     array[CONST.BLEND_MODES.HARD_LIGHT] = new HardLightShader(gl);
+    array[CONST.BLEND_MODES.SOFT_LIGHT] = new SoftLightShader(gl);
 
     return array;
 }
